libs/python/ddd/usecases/shared/job-handler/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "job_handler" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under=80 --cov-report html:'../../../../../../coverage/libs/python/ddd/usecases/shared/job-handler/html' --cov-report xml:'../../../../../../coverage/libs/python/ddd/usecases/shared/job-handler/coverage.xml' --html='../../../../../../reports/libs/python/ddd/usecases/shared/job-handler/unittests/html/index.html' --junitxml='../../../../../../reports/libs/python/ddd/usecases/shared/job-handler/unittests/junit.xml'"

[tool.poetry]
name = "python-ddd-usecases-shared-job-handler"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"

  [[tool.poetry.packages]]
  include = "job_handler"

  [tool.poetry.dependencies]
  python = "^3.10"
  warlock = "^2.0.1"

    [tool.poetry.dependencies.shared-py-log]
    path = "../../../../shared/py-log"
    develop = true

    [tool.poetry.dependencies.python-shared-py-debug]
    path = "../../../../shared/py-debug"
    develop = true

    [tool.poetry.dependencies.python-ddd-dtos-config-vault]
    path = "../../../dtos/config-vault"
    develop = true

    [tool.poetry.dependencies.python-ddd-dtos-events-router]
    path = "../../../dtos/events-router"
    develop = true

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
